pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cmdllm"
version = "0.1.0"
description = "使用自然语言与命令行工具交互的智能接口"
readme = "README.md"
authors = [
    {name = "Zhixiang Pan", email = "pzx553@gmail.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = ["cli", "command-line", "llm", "ai"]
dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "openai>=1.0.0",
]
requires-python = ">=3.6"

[project.urls]
"Homepage" = "https://github.com/yexia553/cmdllm"
"Bug Tracker" = "https://github.com/yexia553/cmdllm/issues"

[project.scripts]
cmdllm = "cmdllm.cli:cli"

[tool.setuptools]
packages = ["cmdllm"] 